requirements.txt
ipwhois == 0.11.0
lxml == 3.3.5
mechanize == 0.2.5
python_nmap == 0.4.3
paramiko == 1.15.2
pygeocoder == 1.2.5
pygeoip == 0.3.2
pythonwhois == 2.4.3
requests == 2.7.0
scapy == 2.3.1
selenium == 2.48.0
setuptools == 19.2
shodan == 1.3.6
Pillow == 3.1.0
PyPDF2 == 1.25.1
beautifulsoup4 == 4.4.1
dnspython == 1.12.0
msgpack_python == 0.4.6
